robodocai/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
pypdf
google-generativeai